backend/requirements.txt
# Python Dependencies

## Core Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1

## Database
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
alembic==1.14.0
pgvector==0.3.6

## Media Processing
Pillow==10.4.0
opencv-python-headless==4.10.0.84
imagehash==4.3.1
ffmpeg-python==0.2.0
pytesseract==0.3.13  # OCR for text detection in images

## Document Processing
unstructured[pdf]==0.18.18
langchain-text-splitters==0.3.11
pdfminer.six>=20251107
python-docx==1.1.2

## ML/Embeddings (CPU-friendly)
sentence-transformers==3.3.1
transformers==4.53.0
torch==2.8.0
onnxruntime==1.23.2
open-clip-torch==2.29.0

## Vision Language Model (VLM)
google-generativeai==0.8.3  # Gemini API SDK

## Job Queue
redis==5.2.0  # Optional, for Redis-backed queue

## Utilities
python-multipart==0.0.20
python-dotenv==1.0.1
httpx==0.28.1
prometheus-client==0.23.1

## Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0

## Development
black==24.10.0
flake8==7.1.1
mypy==1.13.0

